Added Initial Support for `tuple` Data Types.
+ Added initial support for `Tuples`. At the moment, there are no differences at runtime between a tuple and an array.

~ Specified an initial capacity for the values stack and frames stack (to prevent unnecessary reallocation).

+++ src/main.rs
@@ -24,7 +24,7 @@ use virtual_machine::InterpretResult;
 use virtual_machine::VirtualMachine;
 
 // Static things
-static FRAMES_MAX: u16 = 1000;
+static FRAMES_MAX: usize = 1000;
 
 /// The main function
 fn main() {
